Add a Soft Delete Feature to Scheduled Events

+++ app/routers/issuer/bond.py
@@ -120,6 +120,7 @@ from app.model.schema import (
     BulkTransferUploadIdResponse,
     BulkTransferUploadRecordResponse,
     BulkTransferUploadResponse,
+    DeleteScheduledEventQuery,
     GetBatchIssueRedeemResponse,
     GetBatchRegisterPersonalInfoResponse,
     HolderCountResponse,
@@ -1585,6 +1586,7 @@ async def list_all_scheduled_bond_token_update_events(
                 "status": _token_event.status,
                 "data": _token_event.data,
                 "created": created_utc.astimezone(local_tz).isoformat(),
+                "is_soft_deleted": _token_event.is_soft_deleted,
             }
         )
     return json_response(token_events)
@@ -1868,6 +1870,7 @@ async def retrieve_scheduled_bond_token_update_event(
             "status": _token_event.status,
             "data": _token_event.data,
             "created": created_utc.astimezone(local_tz).isoformat(),
+            "is_soft_deleted": _token_event.is_soft_deleted,
         }
     )
 
@@ -1882,6 +1885,7 @@ async def retrieve_scheduled_bond_token_update_event(
 async def delete_scheduled_bond_token_update_event(
     db: DBAsyncSession,
     request: Request,
+    delete_param: Annotated[DeleteScheduledEventQuery, Query()],
     token_address: Annotated[str, Path()],
     scheduled_event_id: Annotated[str, Path()],
     issuer_address: Annotated[str, Header()],
@@ -1936,9 +1940,15 @@ async def delete_scheduled_bond_token_update_event(
         "status": _token_event.status,
         "data": _token_event.data,
         "created": created_utc.astimezone(local_tz).isoformat(),
+        "is_soft_deleted": delete_param.soft_delete,
     }
 
-    await db.delete(_token_event)
+    if delete_param.soft_delete is True:
+        _token_event.is_soft_deleted = True
+        await db.merge(_token_event)
+    else:
+        await db.delete(_token_event)
+
     await db.commit()
 
     return json_response(rtn)
